# docs/openapi.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "重生之凡人修仙传 API",
    "description": "修仙文字游戏后端接口文档，基于《凡人修仙传》世界观。包含玩家注册登录、角色管理、境界突破、地图探索、闭关修炼、GM管理、系统通知、WebSocket实时推送、玩家数据实时同步等完整功能。",
    "version": "2.2.0",
    "contact": {
      "name": "Shell",
      "email": "1592363624@qq.com"
    }
  },
  "servers": [
    {
      "url": "http://localhost:5000",
      "description": "本地开发环境"
    },
    {
      "url": "https://xx.52shell.ltd",
      "description": "生产环境"
    }
  ],
...
  "security": [
    {
      "bearerAuth": []
    }
  ],
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "JWT Token 认证"
      }
    },
    "schemas": {
...
      "Player": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "description": "玩家唯一ID"
          },
          "username": {
            "type": "string",
            "description": "登录账号"
          },
          "nickname": {
            "type": "string",
            "description": "道号/昵称"
          },
          "realm": {
            "type": "string",
            "description": "当前境界"
          },
          "exp": {
            "type": "string",
            "description": "当前修为值"
          },
          "exp_next": {
            "type": "string",
            "description": "下一境界所需修为"
          },
          "exp_cap": {
            "type": "string",
            "description": "修为上限"
          },
          "spirit_root": {
            "type": "object",
            "description": "灵根属性",
            "properties": {
              "type": {
                "type": "string",
                "description": "灵根类型：金/木/水/火/土/雷/冰等"
              },
              "level": {
                "type": "integer",
                "description": "灵根等级 1-10"
              }
            }
          },
          "spirit_stones": {
            "type": "string",
            "description": "灵石数量"
          },
          "age": {
            "type": "integer",
            "description": "当前年龄"
          },
          "lifespan": {
            "type": "object",
            "description": "寿命信息",
            "properties": {
              "current": {
                "type": "integer",
                "description": "当前寿命"
              },
              "max": {
                "type": "integer",
                "description": "最大寿命"
              },
              "percent": {
                "type": "number",
                "description": "寿命百分比"
              }
            }
          },
          "attributes": {
            "type": "object",
            "description": "角色属性",
            "properties": {
              "hp_max": {
                "type": "integer",
                "description": "最大气血"
              },
              "mp_max": {
                "type": "integer",
                "description": "最大灵力，根据境界增长曲线动态计算（炼气、筑基线性增长，金丹及以上使用初期/中期/后期/大圆满阶段命名，采用指数增长模型）"
              },
              "atk": {
                "type": "integer",
                "description": "攻击力"
              },
              "def": {
                "type": "integer",
                "description": "防御力"
              },
              "spd": {
                "type": "integer",
                "description": "速度"
              }
            }
          },
          "is_secluded": {
            "type": "boolean",
            "description": "是否正在闭关"
          },
          "seclusion_end_time": {
            "type": "string",
            "format": "date-time",
            "description": "闭关结束时间"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "description": "创建时间"
          },
          "updated_at": {
            "type": "string",
            "format": "date-time",
            "description": "最后更新时间"
          }
        }
      },
      "Realm": {
        "type": "object",
        "description": "境界配置",
        "properties": {
          "id": {
            "type": "string",
            "description": "境界ID"
          },
          "name": {
            "type": "string",
            "description": "境界名称"
          },
          "rank": {
            "type": "integer",
            "description": "境界等级"
          },
          "description": {
            "type": "string",
            "description": "境界描述"
          },
          "exp_cap": {
            "type": "string",
            "description": "所需修为"
          },
          "base_hp": {
            "type": "integer",
            "description": "基础气血"
          },
          "base_mp": {
            "type": "integer",
            "description": "基础灵力"
          },
          "lifespan_max": {
            "type": "integer",
            "description": "最大寿命"
          }
        }
      },
...
    }
  },
  "paths": {
...
    "/api/player/me": {
      "get": {
        "summary": "获取当前玩家信息",
        "description": "获取玩家基本信息、境界、属性、修为等综合数据",
        "tags": ["Player"],
        "security": [{ "bearerAuth": [] }],
        "responses": {
          "200": {
            "description": "成功",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "code": {
                      "type": "integer",
                      "example": 200
                    },
                    "data": {
                      "$ref": "#/components/schemas/Player"
                    }
                  }
                },
                "examples": {
                  "normal": {
                    "value": {
                      "code": 200,
                      "data": {
                        "id": 1,
                        "username": "hanli",
                        "nickname": "韩天尊",
                        "realm": "炼气13层",
                        "realmInfo": {
                          "name": "炼气13层",
                          "rank": 14,
                          "description": "炼气期巅峰，即将突破筑基"
                        },
                        "exp": "2744000",
                        "exp_cap": "2744000",
                        "exp_progress": 100,
                        "can_breakthrough": true,
                        "spirit_root": {
                          "type": "雷",
                          "level": 9
                        },
                        "spirit_stones": "1500",
                        "age": 52,
                        "lifespan": {
                          "current": 120,
                          "max": 150,
                          "percent": 80
                        },
                        "attributes": {
                          "hp_max": 1200,
                          "mp_max": 800,
                          "atk": 150,
                          "def": 80,
                          "spd": 60
                        },
                        "hp_current": 1200,
                        "hp_max": 1200,
                        "mp_current": 800,
                        "mp_max": 800,
                        "is_secluded": false,
                        "seclusion_end_time": null,
                        "created_at": "2024-05-01T08:00:00.000Z",
                        "updated_at": "2024-05-20T15:30:00.000Z"
                      }
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "未登录或Token无效",
            "content": {
              "application/json": {
                "example": {
                  "code": 401,
                  "message": "未登录或Token已过期"
                }
              }
            }
          }
        }
      },
      "put": {
        "summary": "更新玩家信息",
        "description": "更新玩家基本信息（需要白名单验证）",
        "tags": ["Player"],
        "security": [{ "bearerAuth": [] }],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "nickname": {
                    "type": "string",
                    "description": "新道号"
                  },
                  "hp_current": {
                    "type": "number",
                    "description": "当前气血"
                  },
                  "mp_current": {
                    "type": "number",
                    "description": "当前灵力"
                  },
                  "is_secluded": {
                    "type": "boolean",
                    "description": "是否闭关中"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "更新成功",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "code": {
                      "type": "integer",
                      "example": 200
                    },
                    "message": {
                      "type": "string",
                      "example": "更新成功"
                    },
                    "data": {
                      "$ref": "#/components/schemas/Player"
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/player/attributes": {
      "get": {
        "summary": "获取玩家属性详情",
        "description": "获取玩家的详细属性面板，包括各项战斗属性和加成来源",
        "tags": ["Player"],
        "security": [{ "bearerAuth": [] }],
        "responses": {
          "200": {
            "description": "成功",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "code": {
                      "type": "integer",
                      "example": 200
                    },
                    "data": {
                      "type": "object",
                      "properties": {
                        "base": {
                          "type": "object",
                          "description": "基础属性"
                        },
                        "bonus": {
                          "type": "object",
                          "description": "境界/装备加成"
                        },
                        "final": {
                          "type": "object",
                          "description": "最终属性"
                        }
                      }
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/player/breakthrough": {
      "post": {
        "summary": "执行境界突破",
        "description": "当玩家修为达到当前境界上限时，可以执行突破进入下一境界",
        "tags": ["Player"],
        "security": [{ "bearerAuth": [] }],
        "responses": {
          "200": {
            "description": "突破成功",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "code": {
                      "type": "integer",
                      "example": 200
                    },
                    "message": {
                      "type": "string",
                      "example": "突破成功！当前境界：筑基期"
                    },
                    "data": {
                      "type": "object",
                      "properties": {
                        "newRealm": {
                          "$ref": "#/components/schemas/Realm"
                        }
                      }
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "突破失败（修为不足或已达最高境界）",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "code": {
                      "type": "integer",
                      "example": 400
                    },
                    "message": {
                      "type": "string",
                      "example": "修为不足，需要 100 点修为才能突破"
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
...
  },
...
}